# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Дом в порядке API
  version: 0.1.0
  description: Собственный API MVP. Все маршруты `/api/*`, кроме health, требуют cookie-сессию диспетчера.
servers:
  - url: https://jorwey.ru
    description: Production
paths:
...
  /api/admin/session:
    get:
      security: [{ cookieAuth: [] }]
      summary: Текущая сессия диспетчера
      responses: { '200': { description: Сессия }, '401': { $ref: '#/components/responses/Unauthorized' } }
    post:
      summary: Вход диспетчера
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginRequest' } } } }
      responses: { '200': { description: Сессия создана, headers: { Set-Cookie: { schema: { type: string } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '429': { description: Временная блокировка входа } }
    delete:
      security: [{ cookieAuth: [] }]
      summary: Выход
      responses: { '204': { description: Сессия удалена } }
...
  /api/apartments:
    get:
      security: [{ cookieAuth: [] }]
      summary: Поиск квартир и жителей
      parameters:
        - { name: buildingId, in: query, schema: { type: string, format: uuid } }
        - { name: query, in: query, schema: { type: string, maxLength: 100 } }
        - { name: sort, in: query, schema: { type: string, enum: [apartment, resident] } }
        - { name: withResidents, in: query, schema: { type: boolean } }
      responses: { '200': { description: Список квартир; plaintext индивидуальных кодов не возвращается } }
  /api/apartments/{id}:
    patch:
      security: [{ cookieAuth: [] }]
      summary: Изменить квартиру, жителей, автомобили и индивидуальный код
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ApartmentPatch' } } } }
      responses: { '200': { description: Данные сохранены }, '409': { description: Дубликат номера или автомобиля } }
    delete:
      security: [{ cookieAuth: [] }]
      summary: Удалить квартиру, профиль, автомобили и привязки
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '204': { description: Удалено }, '404': { $ref: '#/components/responses/NotFound' } }
  /api/announcements:
    post:
      security: [{ cookieAuth: [] }]
      summary: Поставить объявление в очередь MAX
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Announcement' } } } }
      responses: { '202': { description: Рассылка поставлена в очередь }, '404': { description: Получатели не найдены } }
  /api/announcements/apartments:
    get:
      security: [{ cookieAuth: [] }]
      summary: Квартиры для адресной рассылки
      responses: { '200': { description: Список квартир с активными жителями } }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: smart_home_admin }
...
  responses:
    Unauthorized: { description: Требуется сессия, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    NotFound: { description: Не найдено, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  schemas:
...
    Error: { type: object, required: [code], properties: { code: { type: string }, requestId: { type: string } } }
    LoginRequest: { type: object, required: [username, password], properties: { username: { type: string }, password: { type: string, minLength: 8 } } }
...
    ApartmentPatch: { type: object, required: [number, residentsCount, people, vehicles], properties: { number: { type: string }, residentsCount: { type: [integer, 'null'] }, inviteCode: { type: [string, 'null'], description: При передаче сохраняется только хеш }, people: { type: array }, vehicles: { type: array } } }
    Announcement: { type: object, required: [audience, text], properties: { audience: { type: string, enum: [all, apartment] }, apartmentId: { type: string, format: uuid }, text: { type: string, minLength: 3, maxLength: 2000 } } }